requirements-dev.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --output-file=requirements-dev.txt requirements/dev.in
#
amqp==2.6.1               # via kombu
appdirs==1.4.4            # via black, virtualenv
attrs==20.1.0             # via pytest
billiard==3.6.3.0         # via celery
black==20.8b1             # via -r requirements/dev.in
celery==4.4.7             # via -r requirements/common.in
cfgv==3.2.0               # via pre-commit
click==7.1.2              # via black, pip-tools
dataclasses==0.7          # via black
distlib==0.3.1            # via virtualenv
django-environ==0.4.5     # via -r requirements/common.in
django==1.11.29           # via -r requirements/common.in, djangorestframework, drf-orjson-renderer
djangorestframework==3.11.1  # via -r requirements/common.in, drf-orjson-renderer
drf-orjson-renderer==1.1.4  # via -r requirements/common.in
factory-boy==3.0.1        # via -r requirements/dev.in
faker==4.1.2              # via factory-boy
filelock==3.0.12          # via virtualenv
flake8-isort==4.0.0       # via -r requirements/dev.in
flake8==3.8.3             # via -r requirements/dev.in, flake8-isort
identify==1.4.29          # via pre-commit
importlib-metadata==1.7.0  # via flake8, kombu, pluggy, pre-commit, pytest, virtualenv
importlib-resources==3.0.0  # via pre-commit, virtualenv
iniconfig==1.0.1          # via pytest
isort==5.4.2              # via flake8-isort
kombu==4.6.11             # via celery
mccabe==0.6.1             # via flake8
more-itertools==8.5.0     # via pytest
mypy-extensions==0.4.3    # via black
nodeenv==1.5.0            # via pre-commit
orjson==3.3.1             # via drf-orjson-renderer
packaging==20.4           # via pytest
pathspec==0.8.0           # via black
pip-tools==5.3.1          # via -r requirements/dev.in
pluggy==0.13.1            # via pytest
pre-commit==2.7.1         # via -r requirements/dev.in
psycopg2==2.8.5           # via -r requirements/common.in
py==1.9.0                 # via pytest
pycodestyle==2.6.0        # via flake8
pyflakes==2.2.0           # via flake8
pyparsing==2.4.7          # via packaging
pytest-django==3.9.0      # via -r requirements/dev.in
pytest==6.0.1             # via -r requirements/dev.in, pytest-django
python-dateutil==2.8.1    # via faker
pytz==2020.1              # via celery, django
pyyaml==5.3.1             # via pre-commit
redis==3.5.3              # via -r requirements/common.in
regex==2020.7.14          # via black
six==1.15.0               # via packaging, pip-tools, python-dateutil, virtualenv
testfixtures==6.14.1      # via flake8-isort
text-unidecode==1.3       # via faker
toml==0.10.1              # via black, pre-commit, pytest
typed-ast==1.4.1          # via black
typing-extensions==3.7.4.3  # via black
vine==1.3.0               # via amqp, celery
virtualenv==20.0.31       # via pre-commit
zipp==3.1.0               # via importlib-metadata, importlib-resources

# The following packages are considered to be unsafe in a requirements file:
# pip
